consolidate nft sending
